requirements.txt
accelerate==1.12.0
anyio==4.12.0
bitsandbytes==0.49.0
certifi==2026.1.4
charset-normalizer==3.4.4
decorator==4.4.2
diffusers==0.36.0
filelock==3.20.2
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
optimum-quanto==0.2.7
packaging==25.0
pillow==11.3.0
proglog==0.1.12
protobuf==6.33.2
psutil==7.2.1
python-dotenv==1.2.1
PyYAML==6.0.3
quanto==0.2.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.20.3
torch==2.9.1
torchaudio==2.9.1
torchsde==0.2.6
torchvision==0.24.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.46.3
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.6.2
zipp==3.23.0
